requirements.txt
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pycrypto==2.6.1
python-dotenv==0.12.0
six==1.14.0
Werkzeug==1.0.0
